# chord-chat/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Chord chat",
    "version": "0.0.1",
    "description": "This is the documentation for chord chat. In this documentation you can view the different routes involved in the craft of chord chat. Chord chat is built with microservices, thus you can view the routes categorized under different services."
  },
  "servers": [
    {
      "url": " http://localhost:3000",
      "description": "Local dev"
    }
  ],
...
  "paths": {
...
    "/create-ad": {
      "post": {
        "tags": [
          "Marketplace-Service"
        ],
        "description": "route to create an ad by a user",
        "requestBody": {
          "description": "This body contains the details to create an ad",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "This contains description for the ad",
                    "example": "Gibson guitar"
                  },
                  "image": {
                    "type": "string",
                    "description": "This contains the url for image in multer storage",
                    "example": "112312babdabsdasd.jpg"
                  },
                  "price": {
                    "type": "string",
                    "description": "This contains the price of the product",
                    "example": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ad created successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "userId": 2,
                  "description": "New guitar",
                  "price": 2000,
                  "image": "fjdashfajofa.jpg",
                  "created_At": "10/04/2024",
                  "Active": true
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Bad request"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/get-ads": {
      "get": {
        "tags": [
          "Marketplace-Service"
        ],
        "description": "route to delete an ad",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "message": "Successfully fetched ads",
                "adDetails": [
                  {
                    "id": 1,
                    "userId": 2,
                    "description": "New guitar",
                    "price": 2000,
                    "image": "fjdashfajofa.jpg",
                    "created_At": "10/04/2024",
                    "Active": true
                  },
                  {
                    "id": 2,
                    "userId": 1,
                    "description": "New drums",
                    "price": 5000,
                    "image": "3313jofa.jpg",
                    "created_At": "11/04/2024",
                    "Active": true
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/view-notifications": {
      "get": {
        "tags": [
          "Notification-Service"
        ],
        "description": "Route to view notifications",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Viewing notifications",
                      "example": "Notifications fetched successfully"
                    },
                    "data": {
                      "type": "object",
                      "example": {
                        "id": 1,
                        "notification_content": {
                          "message": "Welcome to chord chat buds!",
                          "follower_id": -1
                        },
                        "status": "sent",
                        "created_At": "10/04/2024",
                        "generic": true,
                        "user_id": -1
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/send-general-notifications": {
      "post": {
        "tags": [
          "Notification-Service"
        ],
        "description": "Route to send general notifications",
        "requestBody": {
          "description": "This body contains notification to be sent.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notification": {
                    "type": "string",
                    "description": "contains the notification to be sent",
                    "example": "This is a notification to the user"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Sending notifications",
                      "example": "Event generated to notification producer to be recieved from user side to push to notification array to view notifications!"
                    },
                    "data": {
                      "type": "object",
                      "example": {
                        "id": 1,
                        "notification_content": {
                          "message": "This is a notification sent to the user",
                          "follower_id": -1
                        },
                        "status": "sent",
                        "created_At": "10/04/2024",
                        "generic": true,
                        "user_id": -1
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}